Crate load_cargo

Loads a Cargo project into a static instance of analysis, without support for incorporating changes.

Structs

Expander 🔒

LoadCargoConfig

ProjectFolders

SourceRootConfig

Enums

ProcMacroServerChoice

Functions

expander_to_proc_macro 🔒

load_crate_graph_into_db 🔒

load_proc_macro

Load the proc-macros for the given lib path, disabling all expanders whose names are in ignored_macros.

load_workspace

load_workspace_at

load_workspace_into_db

resolve_sub_span 🔒